Add video driver for bcm2835 by Stephen Warren
https://github.com/swarren/u-boot/commit/be33d37f4f906bedd56bc92ec1e08ca710de04ef

The firmware running on the bcm2835 SoC's VideoCore CPU manages the
display controller. Add a simple "LCD" driver that communicates with the
firmware using the property mailbox protocol. This configures the
display and frame-buffer to match whatever physical resolution the
firmware chosen when booting, which is typically the native resolution
of the attached display device, presumably unless otherwise specified
in config.txt on the boot media.

Enable this driver in the Raspberry Pi board configuration.

Signed-off-by: Stephen Warren <swarren@wwwdotorg.org>
---
v3: Re-license to GPL-v2 or later. Fix typo in commit description.
v2: New patch.